+++ actix-session/src/session.rs
@@ -1,6 +1,5 @@
 use std::{
     cell::{Ref, RefCell},
-    collections::HashMap,
     convert::Infallible,
     error::Error as StdError,
     mem,
@@ -16,6 +15,7 @@ use actix_web::{
 use anyhow::Context;
 use derive_more::derive::{Display, From};
 use serde::{de::DeserializeOwned, Serialize};
+use serde_json::{Map, Value};
 
 /// The primary interface to access and modify session state.
 ///
@@ -73,7 +73,7 @@ pub enum SessionStatus {
 
 #[derive(Default)]
 struct SessionInner {
-    state: HashMap<String, String>,
+    state: Map<String, Value>,
     status: SessionStatus,
 }
 
@@ -82,9 +82,9 @@ impl Session {
     ///
     /// It returns an error if it fails to deserialize as `T` the JSON value associated with `key`.
     pub fn get<T: DeserializeOwned>(&self, key: &str) -> Result<Option<T>, SessionGetError> {
-        if let Some(val_str) = self.0.borrow().state.get(key) {
+        if let Some(value) = self.0.borrow().state.get(key) {
             Ok(Some(
-                serde_json::from_str(val_str)
+                serde_json::from_value::<T>(value.clone())
                     .with_context(|| {
                         format!(
                             "Failed to deserialize the JSON-encoded session data attached to key \
@@ -107,8 +107,8 @@ impl Session {
 
     /// Get all raw key-value data from the session.
     ///
-    /// Note that values are JSON encoded.
-    pub fn entries(&self) -> Ref<'_, HashMap<String, String>> {
+    /// Note that values are JSON values.
+    pub fn entries(&self) -> Ref<'_, Map<String, Value>> {
         Ref::map(self.0.borrow(), |inner| &inner.state)
     }
 
@@ -138,7 +138,7 @@ impl Session {
             }
 
             let key = key.into();
-            let val = serde_json::to_string(&value)
+            let val = serde_json::to_value(&value)
                 .with_context(|| {
                     format!(
                         "Failed to serialize the provided `{}` type instance as JSON in order to \
@@ -176,8 +176,8 @@ impl Session {
         let mut inner = self.0.borrow_mut();
         let key_str = key.into();
 
-        if let Some(val_str) = inner.state.get(&key_str) {
-            let value = serde_json::from_str(val_str)
+        if let Some(val) = inner.state.get(&key_str) {
+            let value = serde_json::from_value(val.clone())
                 .with_context(|| {
                     format!(
                         "Failed to deserialize the JSON-encoded session data attached to key \
@@ -187,7 +187,7 @@ impl Session {
                 })
                 .map_err(SessionUpdateError)?;
 
-            let val = serde_json::to_string(&updater(value))
+            let val = serde_json::to_value(updater(value))
                 .with_context(|| {
                     format!(
                         "Failed to serialize the provided `{}` type instance as JSON in order to \
@@ -233,8 +233,8 @@ impl Session {
 
     /// Remove value from the session.
     ///
-    /// If present, the JSON encoded value is returned.
-    pub fn remove(&self, key: &str) -> Option<String> {
+    /// If present, the JSON value is returned.
+    pub fn remove(&self, key: &str) -> Option<Value> {
         let mut inner = self.0.borrow_mut();
 
         if inner.status != SessionStatus::Purged {
@@ -250,10 +250,10 @@ impl Session {
     /// Remove value from the session and deserialize.
     ///
     /// Returns `None` if key was not present in session. Returns `T` if deserialization succeeds,
-    /// otherwise returns un-deserialized JSON string.
-    pub fn remove_as<T: DeserializeOwned>(&self, key: &str) -> Option<Result<T, String>> {
+    /// otherwise returns the raw JSON value.
+    pub fn remove_as<T: DeserializeOwned>(&self, key: &str) -> Option<Result<T, Value>> {
         self.remove(key)
-            .map(|val_str| match serde_json::from_str(&val_str) {
+            .map(|value| match serde_json::from_value::<T>(value.clone()) {
                 Ok(val) => Ok(val),
                 Err(_err) => {
                     tracing::debug!(
@@ -262,7 +262,7 @@ impl Session {
                         std::any::type_name::<T>()
                     );
 
-                    Err(val_str)
+                    Err(value)
                 }
             })
     }
@@ -298,11 +298,11 @@ impl Session {
     /// Adds the given key-value pairs to the session on the request.
     ///
     /// Values that match keys already existing on the session will be overwritten. Values should
-    /// already be JSON serialized.
+    /// already be JSON values.
     #[allow(clippy::needless_pass_by_ref_mut)]
     pub(crate) fn set_session(
         req: &mut ServiceRequest,
-        data: impl IntoIterator<Item = (String, String)>,
+        data: impl IntoIterator<Item = (String, Value)>,
     ) {
         let session = Session::get_session(&mut req.extensions_mut());
         let mut inner = session.0.borrow_mut();
@@ -317,7 +317,7 @@ impl Session {
     #[allow(clippy::needless_pass_by_ref_mut)]
     pub(crate) fn get_changes<B>(
         res: &mut ServiceResponse<B>,
-    ) -> (SessionStatus, HashMap<String, String>) {
+    ) -> (SessionStatus, Map<String, Value>) {
         if let Some(s_impl) = res
             .request()
             .extensions()
@@ -326,7 +326,7 @@ impl Session {
             let state = mem::take(&mut s_impl.borrow_mut().state);
             (s_impl.borrow().status.clone(), state)
         } else {
-            (SessionStatus::Unchanged, HashMap::new())
+            (SessionStatus::Unchanged, Map::new())
         }
     }